[be/fix] Long 으로 변환. 이에 따른 로직 에러 수정

See merge request s08-bigdata-dist-sub2/S08P22A202!227

+++ backend/notice/src/main/java/com/dokidoki/notice/db/entity/AuctionRealtime.java
@@ -11,12 +11,12 @@ import lombok.ToString;
 @NoArgsConstructor
 public class AuctionRealtime {
 
-    private long auctionId;
-    private int highestPrice;
-    private int priceSize;
-    private long sellerId;
+    private Long auctionId;
+    private Long highestPrice;
+    private Long priceSize;
+    private Long sellerId;
 
-    private long productId;
+    private Long productId;
     private String productName;
 
     public static AuctionRealtime from(KafkaAuctionRegisterDTO dto) {
@@ -32,7 +32,7 @@ public class AuctionRealtime {
     }
 
     @Builder
-    public AuctionRealtime(long auctionId, int highestPrice, int priceSize, long sellerId, long productId, String productName) {
+    public AuctionRealtime(Long auctionId, Long highestPrice, Long priceSize, Long sellerId, Long productId, String productName) {
         this.auctionId = auctionId;
         this.highestPrice = highestPrice;
         this.priceSize = priceSize;
@@ -41,11 +41,11 @@ public class AuctionRealtime {
         this.productName = productName;
     }
 
-    public void updatePriceSize(int priceSize) {
+    public void updatePriceSize(Long priceSize) {
         this.priceSize = priceSize;
     }
 
-    public int updateHighestPrice() {
+    public Long updateHighestPrice() {
         this.highestPrice = this.highestPrice + this.priceSize;
         return this.highestPrice;
     }
